Package: Add concurrently package
Concurrently allows the running of multiple scripts at the same time. This is necessary if more than one of the required scripts do not terminate.
One could achieve the same end result by opening multiple terminal tabs and running the required scripts separately, but that would be more time consuming.