Create constants.js
Create a file to hold constants. Add a constant that sets the base api url depending on whether it is local or remote.
Although not quite as good as detecting this automatically, this is a big improvement on changing the request string wherever a fetch occurred. Now, switching to the hosted environment only requires setting isLocal to false before deploying Firebase hosting.