CreateBusiness: rewrite createBusiness function
Import and implement the newly created postFormDataAsJson function.
Prevent default action of form and manually post data.
This successfully creates a new user. Moreover, the user stays on the page as the default action is cancelled. This is better for subsequent flow - signin etc.
Form fields do not get cleared.