CreateBusiness: Redirect after account creation
Create a user state. When account has been created and user signed in, user object is assigned to state.
Wrap form in <div> and add redirect statement that runs when user is truthy.
In future, when setting up restrictions, the user object can be checked for custom claims before logging in.