Helpers: Add validateSharedSignup function
Add a method that validates the data from the part of the form that businesses and individual users share. Namely - phone, email and password.
Use validator package for phone and email. For password, just check length.
Currently uses HTML DOM methods to select the appropriate element. However, this needs to be changed as it doesn't work with React.