Homepage: Implement role fetching function
Implement the getRole function here in the parent rather than HeaderText. The role and name can be passed down as props.
This is necessary as the HeaderText component will be split in 3 parts. It's undesirable to have the role being fetched in each when it's only being used in one. It would lead to excessive API calls.