Calendar: Add event clicking functionality
When the user clicks on a task, they are directed to the 'dayList' view of the day the task takes place. They can then see a list of all the tasks on that date.
Also moved the 'events' props so that all props relating to events are located together for the Calendar component.