Calendar: Change initial view
Change the initial view to custom view 'calendar' as specified in the 'views' prop. The previously implemented 'dayGridMonth' value is a default and was the reason no button was selected at the top of the calendar.
This resolves 2 bugs: a crash when selecting an event on initial load before a button had been pressed. And also the failure to display the time formatting correctly on initial load (since 'calendar' was not loaded right away).