Calendar: Check that row time matches
When checking that correct row has been identified, check that the time is correct too. Had previously only checked the title. This would have led to a false positive if two events had the same names. Now they would have to have the same title and start and end times for a false positive.