helpers: Add auth token to postFormDataAsJson
Check if user exists by attempting to log in. If they do, attach the bearer token to the authorisation header.
This allows the method to be used for POST requests for user and business creation (where no prior user is required) and also for POSTing to the /tasks endpoint (where a user is required).