TaskModal: Extract logic from formattedDate
Extract the logic from formattedDate that decides whether the time should be retained from the date string into a separate function.
This is necessary as lumping it in with the formattedDate function implements the logic even when not desired (such as when saving the current timestamp that should be done regardless if task is all day or not).