Homepage: Do not load TaskModal for staff users
Loading the TaskModal for users that have the staff role caused the api request being denied and an error in the browser console.
An alternative resolution would be to have a conditional for staff in the TaskModal itself, but this would lead to other unnecessary (but valid) calls for tasks etc.