controller: Check that user data has not been improperly altered
Notify the developer via an error message if Auth data or Firestore data for the user has been deleted improperly. This can only happen via the Firebase console to which only the repo admin has access.
User data needs to be deleted from both sources to eliminate the issue. It quietly fails by mismatching first and last names in the userData array if an error is not forced here.