TaskModal: Check data for error
Check the fetched user data for an 'error' field and return the function if present. This prevents an error that gets thrown if the subsequent operations are attempted on the wrong data type.
No need to log the error again as this was already done in the other implementation of getUsers (i.e. in <People/>).