Calendar: Prevent repeated api calls
Add a new boolean state variable called tasksFetched. Use this as a flag to indicate if tasks have been fetched. Use this to fetchTasks only once.
The issue with the previous implementation was that if the user did not have any tasks assigned (e.g. new user), this would lead to an infinite loop of api calls.