Calendar: Track current task id
Add a new ref to track the current task id. It's set when a task is clicked in day or week view (i.e. when the task details are expanded).
The reason a state variable was not used is because they are asynchronous unlike ref. This means that any state saved is only accessible after the eventClicked function has run. It is needed before however, leading to state being one 'step' behind if implemented.