Calendar: Add taskColour function for displayInfo
Add a helper function that determines what class name the wrapper tag for task info should have. This is the same as was previously implemented in the ternary operator, except it allows for the 'overdue' name.
Also add the 'overdue' class name to the CSS file with appropriate red colouring.