Fix error resulting from Node v18 being used
Was getting an error as outlined in SO linked below. This was due to upgrading Node from v16 to v18 locally.
This solution will do for now but is not very secure. Forcing an audit fix (the recommended approach) caused breakages so is not a solution without removing some offending packages first.
https://stackoverflow.com/questions/69692842/error-message-error0308010cdigital-envelope-routinesunsupported